Consolidate face tests into single test file
Move TestDeleteFace and TestReassignFace from test_faces_delete_reassign.py
into test_faces.py so all face endpoint tests live in one file matching the
production module.
